module docstrings, some more refactoring
