+++ inputremapper/gui/components/common.py
@@ -18,17 +18,26 @@
 # You should have received a copy of the GNU General Public License
 # along with input-remapper.  If not, see <https://www.gnu.org/licenses/>.
 
+
+"""Components used in multiple places."""
+
+
 from __future__ import annotations
 
 from gi.repository import Gtk
 
 from typing import Optional
 
+from inputremapper.configs.mapping import MappingData
+from inputremapper.event_combination import EventCombination
+
 from inputremapper.gui.controller import Controller
-from inputremapper.gui.message_broker import MessageBroker
+from inputremapper.gui.gettext import _
+from inputremapper.gui.message_broker import MessageBroker, MessageType, UserConfirmRequest, PresetData, GroupData
 from inputremapper.gui.utils import HandlerDisabled
 
 
+# TODO test
 class FlowBoxEntry(Gtk.ToggleButton):
     """A device that can be selected in the GUI.
 
@@ -90,6 +99,7 @@ class FlowBoxEntry(Gtk.ToggleButton):
             self.set_active(active)
 
 
+# TODO test
 class FlowBoxWrapper:
     """A wrapper for a flowbox that contains FlowBoxEntry widgets."""
 
@@ -98,3 +108,116 @@ class FlowBoxWrapper:
         for child in self._gui.get_children():
             flow_box_entry: FlowBoxEntry = child.get_children()[0]
             flow_box_entry.show_active(flow_box_entry.name == name)
+
+
+class ConfirmCancelDialog:
+    """the dialog shown to the user to query a confirm or cancel action form the user"""
+
+    def __init__(
+        self,
+        message_broker: MessageBroker,
+        controller: Controller,
+        window: Gtk.Window,
+    ):
+        self._message_broker = message_broker
+        self._controller = controller
+        self.window = window
+
+        self._message_broker.subscribe(
+            MessageType.user_confirm_request, self._on_user_confirm_request
+        )
+
+    def _on_user_confirm_request(self, msg: UserConfirmRequest):
+        # if the message contains a line-break, use the first chunk for the primary
+        # message, and the rest for the secondary message.
+        chunks = msg.msg.split("\n")
+        primary = chunks[0]
+        secondary = " ".join(chunks[1:])
+
+        message_dialog = Gtk.MessageDialog(
+            self.window,
+            Gtk.DialogFlags.MODAL | Gtk.DialogFlags.DESTROY_WITH_PARENT,
+            Gtk.MessageType.QUESTION,
+            Gtk.ButtonsType.NONE,
+            primary,
+        )
+
+        if secondary:
+            message_dialog.format_secondary_text(secondary)
+
+        message_dialog.add_button("Cancel", Gtk.ResponseType.CANCEL)
+
+        confirm_button = message_dialog.add_button("Confirm", Gtk.ResponseType.ACCEPT)
+        confirm_button.get_style_context().add_class(Gtk.STYLE_CLASS_DESTRUCTIVE_ACTION)
+
+        response = message_dialog.run()
+        msg.respond(response == Gtk.ResponseType.ACCEPT)
+
+        message_dialog.hide()
+
+
+
+# TODO test
+class Breadcrumbs:
+    """Writes a breadcrumbs string into a given label."""
+
+    def __init__(
+        self,
+        message_broker: MessageBroker,
+        label: Gtk.Label,
+        show_device_group: bool = False,
+        show_preset: bool = False,
+        show_mapping: bool = False,
+    ):
+        self._message_broker = message_broker
+        self._gui = label
+        self._connect_message_listener()
+
+        self.show_device_group = show_device_group
+        self.show_preset = show_preset
+        self.show_mapping = show_mapping
+
+        self._group_key: str = ""
+        self._preset_name: str = ""
+        self._mapping_name: str = ""
+
+        label.set_max_width_chars(50)
+        label.set_line_wrap(True)
+        label.set_line_wrap_mode(2)
+
+    def _connect_message_listener(self):
+        self._message_broker.subscribe(MessageType.group, self._on_group_changed)
+        self._message_broker.subscribe(MessageType.preset, self._on_preset_changed)
+        self._message_broker.subscribe(MessageType.mapping, self._on_mapping_changed)
+
+    def _on_preset_changed(self, data: PresetData):
+        self._preset_name = data.name or ""
+        self._render()
+
+    def _on_group_changed(self, data: GroupData):
+        self._group_key = data.group_key
+        self._render()
+
+    def _on_mapping_changed(self, mapping: MappingData):
+        if mapping.name:
+            self._mapping_name = mapping.name
+        elif mapping.event_combination != EventCombination.empty_combination():
+            self._mapping_name = mapping.event_combination.beautify()
+        else:
+            self._mapping_name = _("empty mapping")
+
+        self._render()
+
+    def _render(self):
+        label = []
+
+        if self.show_device_group:
+            label.append(self._group_key)
+
+        if self.show_preset:
+            label.append(self._preset_name)
+
+        if self.show_mapping:
+            label.append(self._mapping_name)
+
+        self._gui.set_label("  /  ".join(label))
